refactor(*): move common code to core and tidy up imports
