+++ crates/go/src/path.rs
@@ -1,5 +1,9 @@
 use heck::ToSnakeCase;
-use wit_bindgen_core::{wit_parser::{WorldKey, Resolve, PackageId}, Direction};
+use wit_bindgen_core::{
+    name_package_module,
+    wit_parser::{Resolve, WorldKey},
+    Direction,
+};
 
 pub(crate) trait GoPath {
     fn to_path(&self, resolve: &Resolve, direction: Direction) -> Vec<String>;
@@ -11,62 +15,17 @@ impl GoPath for WorldKey {
         if matches!(direction, Direction::Export) {
             path.push("exports".to_string());
         }
-        match self {   
+        match self {
             WorldKey::Name(n) => path.push(n.to_snake_case()),
             WorldKey::Interface(id) => {
-            let iface = &resolve.interfaces[*id];
-            let pkg = iface.package.unwrap();
-            let pkgname = resolve.packages[pkg].name.clone();
-            path.push(pkgname.namespace.to_snake_case());
-            path.push(name_package_module(resolve, pkg));
-            path.push(iface.name.as_ref().unwrap().to_snake_case());
+                let iface = &resolve.interfaces[*id];
+                let pkg = iface.package.unwrap();
+                let pkgname = resolve.packages[pkg].name.clone();
+                path.push(pkgname.namespace.to_snake_case());
+                path.push(name_package_module(resolve, pkg));
+                path.push(iface.name.as_ref().unwrap().to_snake_case());
             }
         }
         path
     }
-}
-
-/// If the package `id` is the only package with its namespace/name combo
-/// then pass through the name unmodified. If, however, there are multiple
-/// versions of this package then the package module is going to get version
-/// information.
-pub fn name_package_module(resolve: &Resolve, id: PackageId) -> String {
-    let pkg = &resolve.packages[id];
-    let versions_with_same_name = resolve
-        .packages
-        .iter()
-        .filter_map(|(_, p)| {
-            if p.name.namespace == pkg.name.namespace && p.name.name == pkg.name.name {
-                Some(&p.name.version)
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-    let base = pkg.name.name.to_snake_case();
-    if versions_with_same_name.len() == 1 {
-        return base;
-    }
-
-    let version = match &pkg.name.version {
-        Some(version) => version,
-        // If this package didn't have a version then don't mangle its name
-        // and other packages with the same name but with versions present
-        // will have their names mangled.
-        None => return base,
-    };
-
-    // Here there's multiple packages with the same name that differ only in
-    // version, so the version needs to be mangled into the Rust module name
-    // that we're generating. This in theory could look at all of
-    // `versions_with_same_name` and produce a minimal diff, e.g. for 0.1.0
-    // and 0.2.0 this could generate "foo1" and "foo2", but for now
-    // a simpler path is chosen to generate "foo0_1_0" and "foo0_2_0".
-    let version = version
-        .to_string()
-        .replace('.', "_")
-        .replace('-', "_")
-        .replace('+', "_")
-        .to_snake_case();
-    format!("{base}{version}")
 }
